The cell distribution triggered the unmapped patch field error and caused float point exceptions
The reason was that the processor patches were moved into an non-empty boundary patch. Therefore the directmapper was unable to map the entire field as the field was bigger than expected. The non-mapable values where NOT initialized and could cause an FPE

The solution is to create a temporary calculatedFvPatchField "oldInternalFace" and move the processor patches in that field. This BC sets its values if not properly mapped with patchInternalField